--- src/app/register/page.tsx
@@ -1,12 +1,0 @@
-'use client';
-import RegisterForm from '@/components/organisms/RegisterForm';
-
-function Register() {
-  return (
-    <div>
-      <RegisterForm />
-    </div>
-  );
-}
-
-export default Register;
